meho-claude/pyproject.toml
[project]
name = "meho-claude"
version = "0.1.0"
description = "MEHO for Claude - Terminal DevOps Intelligence Engine"
requires-python = ">=3.13"
authors = [{name = "evoila"}]

dependencies = [
    # CLI
    "typer>=0.24.0",
    "rich>=14.0.0",

    # Data validation & config
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Security
    "cryptography>=41.0.0",

    # HTTP client
    "httpx>=0.28.1",

    # OpenAPI parsing
    "prance>=25.4.8.0",
    "openapi-spec-validator>=0.7.0",
    "openapi-pydantic>=0.5.1",

    # Data & search
    "duckdb>=1.4.4",
    "chromadb>=1.5.2",
    "sentence-transformers>=3.0.0",

    # OAuth2
    "authlib>=1.6.6",

    # Serialization
    "msgspec>=0.20.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",

    # Logging
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",

    # Kubernetes
    "kubernetes-asyncio>=35.0.1",

    # VMware
    "pyvmomi>=9.0.0.0",

    # SOAP/WSDL
    "zeep>=4.3.2",

    # Proxmox
    "proxmoxer>=2.2.0",
    "requests>=2.31.0",

    # GCP
    "google-cloud-compute>=1.44.0",
    "google-cloud-container>=2.63.0",
    "google-cloud-monitoring>=2.29.1",
    "google-api-python-client>=2.0.0",

    # Knowledge ingestion
    "pymupdf4llm>=0.0.17",
    "markdownify>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.1.7",
    "mypy>=1.7.0",
]

[project.scripts]
meho = "meho_claude.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meho_claude"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
